+++ src/main/java/fuzs/stylisheffects/client/handler/EffectScreenHandler.java
@@ -7,7 +7,6 @@ import fuzs.stylisheffects.client.gui.effects.VanillaEffectRenderer;
 import fuzs.stylisheffects.config.ClientConfig;
 import fuzs.stylisheffects.mixin.client.accessor.ContainerAccessor;
 import fuzs.stylisheffects.mixin.client.accessor.DisplayEffectsScreenAccessor;
-import net.minecraft.client.MainWindow;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.DisplayEffectsScreen;
 import net.minecraft.client.gui.recipebook.IRecipeShownListener;
@@ -25,45 +24,23 @@ import net.minecraftforge.client.event.RenderGameOverlayEvent.ElementType;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import javax.annotation.Nullable;
 import java.util.function.Consumer;
-import java.util.function.Function;
 
 public class EffectScreenHandler {
-    @Nullable
-    public static AbstractEffectRenderer inventoryRenderer;
-    @Nullable
-    public static AbstractEffectRenderer hudRenderer;
+    private static AbstractEffectRenderer hudRenderer;
 
     public static void createHudRenderer() {
-        Consumer<AbstractEffectRenderer> setScreenDimensions = renderer -> {
-            Minecraft minecraft = Minecraft.getInstance();
-            final ClientConfig.ScreenSide screenSide = StylishEffects.CONFIG.client().hudRenderer().screenSide;
-            final MainWindow window = minecraft.getWindow();
-            renderer.setScreenDimensions(minecraft.gui, window.getGuiScaledWidth(), window.getGuiScaledHeight(), screenSide.right() ? window.getGuiScaledWidth() : 0, 0, screenSide);
-        };
-        hudRenderer = createEffectRenderer(StylishEffects.CONFIG.client().hudRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.HUD, setScreenDimensions);
+        hudRenderer = createEffectRenderer(StylishEffects.CONFIG.client().hudRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.HUD);
     }
 
-    @Nullable
-    private static AbstractEffectRenderer createEffectRenderer(ClientConfig.EffectRenderer rendererType, AbstractEffectRenderer.EffectRendererType effectRendererType, Consumer<AbstractEffectRenderer> setScreenDimensions) {
+    private static AbstractEffectRenderer createEffectRenderer(ClientConfig.EffectRenderer rendererType, AbstractEffectRenderer.EffectRendererType effectRendererType) {
         switch (rendererType) {
             case VANILLA:
-                return createRendererOrFallback(effectRendererType, VanillaEffectRenderer::new, setScreenDimensions);
+                return new VanillaEffectRenderer(effectRendererType);
             case COMPACT:
-                return createRendererOrFallback(effectRendererType, CompactEffectRenderer::new, setScreenDimensions);
+                return new CompactEffectRenderer(effectRendererType);
         }
-        return null;
-    }
-
-    private static AbstractEffectRenderer createRendererOrFallback(AbstractEffectRenderer.EffectRendererType type, Function<AbstractEffectRenderer.EffectRendererType, AbstractEffectRenderer> effectRendererFactory, Consumer<AbstractEffectRenderer> setScreenDimensions) {
-        AbstractEffectRenderer renderer = effectRendererFactory.apply(type);
-        setScreenDimensions.accept(renderer);
-        while (!renderer.isValid()) {
-            renderer = renderer.getFallbackRenderer().apply(type);
-            setScreenDimensions.accept(renderer);
-        }
-        return renderer;
+        throw new IllegalStateException("unreachable statement");
     }
 
     @SubscribeEvent
@@ -84,40 +61,22 @@ public class EffectScreenHandler {
     }
 
     @SubscribeEvent
-    public void onInitGuiPost(final GuiScreenEvent.InitGuiEvent.Post evt) {
-        inventoryRenderer = null;
-    }
-
-    private void extracted(Screen screen) {
-        if (inventoryRenderer != null) return;
-        if (supportsEffectsDisplay(screen)) {
-            Consumer<AbstractEffectRenderer> setScreenDimensions = renderer -> {
-                ContainerScreen<?> containerScreen = (ContainerScreen<?>) screen;
-                final ClientConfig.ScreenSide screenSide = StylishEffects.CONFIG.client().inventoryRenderer().screenSide;
-                renderer.setScreenDimensions(containerScreen, !screenSide.right() ? containerScreen.getGuiLeft() : containerScreen.width - (containerScreen.getGuiLeft() + containerScreen.getXSize()), containerScreen.getYSize(), !screenSide.right() ? containerScreen.getGuiLeft() : containerScreen.getGuiLeft() + containerScreen.getXSize(), containerScreen.getGuiTop(), screenSide);
-            };
-            inventoryRenderer = createEffectRenderer(StylishEffects.CONFIG.client().inventoryRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.INVENTORY, setScreenDimensions);
-        }
-        inventoryRenderer = AbstractEffectRenderer.EMPTY;
-    }
-
-    @SubscribeEvent
     public void onPotionShift(final GuiScreenEvent.PotionShiftEvent evt) {
         evt.setCanceled(true);
     }
 
     @SubscribeEvent
     public void onRenderGameOverlayText(final RenderGameOverlayEvent.Text evt) {
-        // use this event so potion icons are drawn behind debug menu as in vanilla
-        // field may get changed during config reload from different thread
-        final AbstractEffectRenderer hudRenderer = EffectScreenHandler.hudRenderer;
-        if (hudRenderer != null) {
-            final Minecraft minecraft = Minecraft.getInstance();
-            if (minecraft.screen == null || !supportsEffectsDisplay(minecraft.screen)) {
-                hudRenderer.setActiveEffects(minecraft.player.getActiveEffects());
-                if (hudRenderer.isActive()) {
-                    hudRenderer.renderEffects(evt.getMatrixStack(), minecraft);
-                }
+        final Minecraft minecraft = Minecraft.getInstance();
+        if (minecraft.screen == null || !supportsEffectsDisplay(minecraft.screen)) {
+            // use this event so potion icons are drawn behind debug menu as in vanilla
+            // field may get changed during config reload from different thread
+            final AbstractEffectRenderer hudRenderer = EffectScreenHandler.hudRenderer;
+            hudRenderer.setActiveEffects(minecraft.player.getActiveEffects());
+            if (hudRenderer.isActive()) {
+                final ClientConfig.ScreenSide screenSide = StylishEffects.CONFIG.client().hudRenderer().screenSide;
+                hudRenderer.setScreenDimensions(minecraft.gui, evt.getWindow().getGuiScaledWidth(), evt.getWindow().getGuiScaledHeight(), screenSide.right() ? evt.getWindow().getGuiScaledWidth() : 0, 0, screenSide);
+                hudRenderer.renderEffects(evt.getMatrixStack(), minecraft);
             }
         }
     }
@@ -125,10 +84,9 @@ public class EffectScreenHandler {
     @SubscribeEvent
     public void onDrawScreenPost(final GuiScreenEvent.DrawScreenEvent.Post evt) {
         // field may get changed during config reload from different thread
-        final AbstractEffectRenderer inventoryRenderer = EffectScreenHandler.inventoryRenderer;
+        final AbstractEffectRenderer inventoryRenderer = createRendererOrFallback(evt.getGui());
         if (inventoryRenderer != null) {
             final Minecraft minecraft = evt.getGui().getMinecraft();
-            inventoryRenderer.setActiveEffects(minecraft.player.getActiveEffects());
             if (inventoryRenderer.isActive()) {
                 inventoryRenderer.renderEffects(evt.getMatrixStack(), minecraft);
                 inventoryRenderer.getHoveredEffectTooltip(evt.getMouseX(), evt.getMouseY()).ifPresent(tooltip -> {
@@ -158,5 +116,24 @@ public class EffectScreenHandler {
             return true;
         }
         return false;
+    }
+
+    public static AbstractEffectRenderer createRendererOrFallback(Screen screen) {
+        if (supportsEffectsDisplay(screen)) {
+            final ContainerScreen<?> containerScreen = (ContainerScreen<?>) screen;
+            final ClientConfig.ScreenSide screenSide = StylishEffects.CONFIG.client().inventoryRenderer().screenSide;
+            Consumer<AbstractEffectRenderer> setScreenDimensions = renderer -> {
+                renderer.setScreenDimensions(containerScreen, !screenSide.right() ? containerScreen.getGuiLeft() : containerScreen.width - (containerScreen.getGuiLeft() + containerScreen.getXSize()), containerScreen.getYSize(), !screenSide.right() ? containerScreen.getGuiLeft() : containerScreen.getGuiLeft() + containerScreen.getXSize(), containerScreen.getGuiTop(), screenSide);
+            };
+            AbstractEffectRenderer renderer = createEffectRenderer(StylishEffects.CONFIG.client().inventoryRenderer().rendererType, AbstractEffectRenderer.EffectRendererType.INVENTORY);
+            setScreenDimensions.accept(renderer);
+            while (!renderer.isValid()) {
+                renderer = renderer.getFallbackRenderer().apply(AbstractEffectRenderer.EffectRendererType.INVENTORY);
+                setScreenDimensions.accept(renderer);
+            }
+            renderer.setActiveEffects(screen.getMinecraft().player.getActiveEffects());
+            return renderer;
+        }
+        return null;
     }
 }