+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v1.2.0-1.16.5] - 2021-12-21
+### Added
+- Added fallback behavior for effect renderers in case not enough screen space is available
+- Added screen border distance option
+### Changed
+- Hud renderer no longer shows when effect renderer is rendered inside of a screen (like 1.18)
+- Separated screen space options for x- and y-axis
+
 ## [v1.1.1-1.16.5] - 2021-12-17
 ### Added
 - Added a config option for how the vanilla `hideParticles` flag should be respected